layouts: Enhance licenses pages and readfile shortcode readFile usage
For the licenses pages readFile usage, and for the readfile shortcode
(now theme-readfile shortcode), first look for the file as specified
in the shortcode or the licenses directory of the site root, and if
that fails look also in the same place but in the themes/oldnew-mashshup
directory.  This allows to avoid symlinks which confuse hugo's change
detection.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>