layouts assets: Fix potental screen ficker by removing auto font resize
The automatic font resize / scaling can result in frequent screen redraws which can
result in flicker.  Flicker is an accessibility issue as it can trigger an epileptic
seizure in some people.  Therefore remove that function.

With CSS3 media queries we can get font resizing based on screen
resolution as well, without Javascript or as much likeliehood of
screen flicker during resize.  It's not a granular, smooth, but
likely no one will miss it.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>