Fix for unbreakable line in sidebar
When there are unbreakable lines in the sidebar, the sidebar doesn't
beside the main content.  We fix this by setting a maximum width for
the sidebar.  We also set a minimum width so that on small screen the
sidebar remains useful, although below the main content instead of
beside it.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>