treewide: Add site params for colour config
Site enable the use of .Param "xxxx" in a base SCSS file so that can
set our colours via .Site.Params (i.e the params section of the
config file (e.g. in config.toml)).  We also document the params.

The end goal is to have the majority of the theme configurable via
documented configuration parameters.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>