assets scss: Fix menubar sub-menus not displaying
In fixing the case of menu items with names that are too long, we broke
the sub-menus from displaying.  We fix that now by moving the overflow
hiding/ellipsis to the menu item wrapper instead of on the list item
itself.  This allows overflow which allows the sub-menu to be seen when
it is positioned on a different z-index and is therefore outside the
regular flow.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>