treewide: reorganize menu code; fix accessibility
This ended up pulling in a colour scheme change for most of the site as
well, as it wasn't possible to sanely change just the side navigation
bar (part of the menu reorganization).  So this commit reorganize the
menu generation but also alters the colour scheme for the site to be
high contrast.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>