treewide: Update syntax highlighting and tweak
We move the syntax highlighting for Hugo into the new Hugo template
based SCSS, and make some tweaks along the way.  This includes managing
the wrapper 'highlight' div that allows us to keep the 'pre' element
from interfering with the page layout (due to un-wrapped 'pre' elements
not respecting the width of the containing block, unless overflow is
managed, not just the default of 'visible').

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>